Rename CosmicCylinder->CosmicDistance and ScatterDCA->ScatterClosestApproach
