+++ sbnobj/Common/Reco/ScatterClosestApproach.cc
@@ -1,0 +1,1 @@
+#include "sbnobj/Common/Reco/ScatterClosestApproach.h"
