--- sbnobj/Common/Reco/ScatterDCA.cc
@@ -1,1 +1,0 @@
-#include "sbnobj/Common/Reco/ScatterDCA.h"
