+++ sbnobj/Common/Reco/classes.h
@@ -18,7 +18,7 @@
 
 #include "sbnobj/Common/Reco/MVAPID.h"
 #include "sbnobj/Common/Reco/MergedTrackInfo.hh"
-#include "sbnobj/Common/Reco/ScatterDCA.h"
+#include "sbnobj/Common/Reco/ScatterClosestApproach.h"
 #include "sbnobj/Common/Reco/StoppingChi2Fit.h"
 #include "sbnobj/Common/Reco/Stub.h"
 #include "sbnobj/Common/Reco/VertexHit.h"
